+++ doc/changelog.rst
@@ -49,6 +49,7 @@ Core
 Admin
 ~~~~~
 
+- Fix bug: Filter category parent choices based on current shop
 - Add middleware to select and set the current shop in the request
 - Breaking change: add optional shop parameter in ``get_model_url`` method of
   admin module